pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hikvision-viewer"
version = "0.1.0"
description = "PyQt6 multi-stream RTSP viewer for IP cameras"
requires-python = ">=3.10"
dependencies = [
  "PyQt6>=6.6.0",
  "PyYAML>=6.0.1",
  "cryptography>=42.0.0",
  "keyring>=25.0.0",
  "python-dotenv>=1.0.0",
  "python-mpv>=1.0.8",
]

[project.scripts]
hikvision-viewer = "hikvision_viewer.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hikvision_viewer*"]

[tool.setuptools.data-files]
"share/applications" = ["hikvision-viewer.desktop"]
